YH-731: decomposed the structure and removed the excess
