+++ src/entities/collection/index.ts
@@ -9,3 +9,4 @@ export * from './api/collectionApi';
 
 export { ChooseCollectionAccess } from './ui/ChooseCollectionAccess/ChooseCollectionAccess';
 export { CollectionPreview } from './ui/CollectionPreview/CollectionPreview';
+export { PreviewCollectionsItemSkeleton } from './ui/PreviewCollectionItem/PreviewCollectionsItem.skeleton';
