+++ src/entities/collection/ui/PreviewCollectionItem/PreviewCollectionsItem.skeleton.tsx
@@ -1,5 +1,3 @@
-import classNames from 'classnames';
-
 import { CollectionParamSkeleton } from '@/shared/ui/CollectionParam';
 import { Flex } from '@/shared/ui/Flex';
 import { ImageWithWrapperSkeleton } from '@/shared/ui/ImageWithWrapper';
@@ -12,19 +10,14 @@ export const PreviewCollectionsItemSkeleton = () => {
 		<Flex direction="row" componentType="li" gap="8" className={styles.wrapper}>
 			<ImageWithWrapperSkeleton className={styles['image-wrapper']} />
 			<Flex gap="24" className={styles['wrapper-content']}>
-				<Flex direction="row" gap="20" className={styles.header}>
+				<Flex direction="row" gap="20">
 					<CollectionParamSkeleton />
 				</Flex>
 
-				<TextSkeleton
-					variant="body3-accent"
-					width={280}
-					className={classNames(styles.title, styles['title-skeleton'])}
-				/>
-
+				<TextSkeleton variant="body3-accent" width={280} className={styles['title-skeleton']} />
 				<TextSkeleton variant={'body3-accent'} width={40} />
 
-				<Flex direction="row" gap="20" className={styles.params}>
+				<Flex direction="row" gap="20">
 					<TextSkeleton variant={'body3-accent'} width={50} />
 					<TextSkeleton variant={'body3-accent'} width={50} />
 				</Flex>
