+++ src/widgets/interview/PreviewCollectionsList/index.ts
