+++ src/widgets/interview/PreviewCollectionsList/ui/PreviewCollectionsList/PreviewCollectionsList.module.css
